Fix five subtle correctness bugs found in code review
- Exclude forfeit matchups (gross=0) from handicap score history
- Pass weekNumber to weekly scores preview to exclude current week
- Add missing fields to updateSeason Zod schema (year, dates, weeks)
- Use parseFloat instead of parseInt for manual handicap inputs
- Cap manual handicaps in weekly scores to league min/max

+++ src/app/league/[slug]/admin/components/MatchupsTab.tsx
@@ -639,7 +639,7 @@ export default function MatchupsTab({
                       <input
                         type="number"
                         value={teamAHandicapManual}
-                        onChange={(e) => setTeamAHandicapManual(e.target.value ? parseInt(e.target.value) : "")}
+                        onChange={(e) => setTeamAHandicapManual(e.target.value ? parseFloat(e.target.value) : "")}
                         className="w-full pencil-input"
                       />
                     </div>
@@ -715,7 +715,7 @@ export default function MatchupsTab({
                       <input
                         type="number"
                         value={teamBHandicapManual}
-                        onChange={(e) => setTeamBHandicapManual(e.target.value ? parseInt(e.target.value) : "")}
+                        onChange={(e) => setTeamBHandicapManual(e.target.value ? parseFloat(e.target.value) : "")}
                         className="w-full pencil-input"
                       />
                     </div>
